+++ pwncat/modules/linux/implant/pam.py
@@ -2,8 +2,7 @@
 import io
 import hashlib
 from subprocess import CalledProcessError
-
-import pkg_resources
+from importlib.resources import files as _pkg_files
 
 import pwncat
 from pwncat.facts import Implant, CreatedFile
@@ -97,7 +96,7 @@ class Module(ImplantModule):
             raise ModuleFailed("only one pam implant may be installed at a time")
 
         yield Status("loading pam module source code")
-        with open(pkg_resources.resource_filename("pwncat", "data/pam.c"), "r") as filp:
+        with open(str(_pkg_files("pwncat").joinpath("data/pam.c")), "r") as filp:
             sneaky_source = filp.read()
 
         yield Status("checking selinux state")