+++ pwncat/modules/windows/powersploit.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python3
 import os.path
-
-import pkg_resources
+from importlib.resources import files as _pkg_files
 
 import pwncat
 from pwncat.modules import Result, Status, Argument, BaseModule, ModuleFailed
@@ -109,8 +108,8 @@ class Module(BaseModule):
         for url in self.MODULES[group]:
             yield Status(f"loading {url.split('/')[-1]}")
 
-            path = pkg_resources.resource_filename(
-                "pwncat", os.path.join("data/PowerSploit", url)
+            path = str(
+                _pkg_files("pwncat").joinpath(os.path.join("data/PowerSploit", url))
             )
 
             try: