+++ pwncat/platform/linux.py
@@ -23,8 +23,7 @@ import subprocess
 from io import TextIOWrapper, BufferedIOBase, UnsupportedOperation
 from typing import List, Union, BinaryIO, Optional, Generator
 from subprocess import TimeoutExpired, CalledProcessError
-
-import pkg_resources
+from importlib.resources import files as _pkg_files
 
 import pwncat
 import pwncat.channel
@@ -579,7 +578,7 @@ class Linux(Platform):
         # Load a GTFOBins database to assist in common operations
         # without relying on specific binaries being available.
         self.gtfo = GTFOBins(
-            pkg_resources.resource_filename("pwncat", "data/gtfobins.json"), self.which
+            str(_pkg_files("pwncat").joinpath("data/gtfobins.json")), self.which
         )
 
         # Ensure history is disabled